+++ nodeconductor/structure/models.py
@@ -1,9 +1,12 @@
 from __future__ import unicode_literals
 
-import yaml
+import datetime
 import itertools
+import yaml
 
 from django.apps import apps
+from django.core.cache import cache
+from django.conf import settings
 from django.contrib.auth import get_user_model
 from django.contrib.auth.models import Group
 from django.contrib.contenttypes.fields import GenericForeignKey
@@ -22,10 +25,11 @@ from model_utils import FieldTracker
 from model_utils.models import TimeStampedModel
 from model_utils.fields import AutoCreatedField
 from taggit.managers import TaggableManager
-
+import pyvat
 
 from nodeconductor.core import fields as core_fields
 from nodeconductor.core import models as core_models
+from nodeconductor.core import utils as core_utils
 from nodeconductor.core.models import CoordinatesMixin, AbstractFieldTracker
 from nodeconductor.core.tasks import send_task
 from nodeconductor.monitoring.models import MonitoringModelMixin
@@ -80,9 +84,51 @@ class StructureModel(models.Model):
             "'%s' object has no attribute '%s'" % (self._meta.object_name, name))
 
 
+class StructureLoggableMixin(LoggableMixin):
+
+    @classmethod
+    def get_permitted_objects_uuids(cls, user):
+        """
+        Return query dictionary to search objects available to user.
+        """
+        uuids = filter_queryset_for_user(cls.objects.all(), user).values_list('uuid', flat=True)
+        key = core_utils.camel_case_to_underscore(cls.__name__) + '_uuid'
+        return {key: uuids}
+
+
+class TagMixin(models.Model):
+    """
+    Add tags field and manage cache for tags.
+    """
+    class Meta:
+        abstract = True
+
+    tags = TaggableManager(related_name='+', blank=True)
+
+    def get_tags(self):
+        key = self._get_tag_cache_key()
+        tags = cache.get(key)
+        if tags is None:
+            tags = list(self.tags.all().values_list('name', flat=True))
+            cache.set(key, tags)
+        return tags
+
+    def clean_tag_cache(self):
+        key = self._get_tag_cache_key()
+        cache.delete(key)
+
+    def _get_tag_cache_key(self):
+        return 'tags:%s' % core_utils.serialize_instance(self)
+
+
+class VATException(Exception):
+    pass
+
+
 class VATMixin(models.Model):
     """
-    Add VAT number field and check results from EU VAT Information Exchange System.
+    Add country, VAT number fields and check results from EU VAT Information Exchange System.
+    Allows to compute VAT charge rate.
     """
     class Meta(object):
         abstract = True
@@ -92,6 +138,30 @@ class VATMixin(models.Model):
                                 help_text='Optional business name retrieved for the VAT number.')
     vat_address = models.CharField(max_length=255, blank=True,
                                    help_text='Optional business address retrieved for the VAT number.')
+
+    is_company = models.BooleanField(default=False, help_text="Is company or private person")
+    country = core_fields.CountryField(blank=True)
+
+    def get_vat_rate(self):
+        charge = self.get_vat_charge()
+        if charge.action == pyvat.VatChargeAction.charge:
+            return charge.rate
+        # Return None, if reverse_charge or no_charge action is applied
+
+    def get_vat_charge(self):
+        if not self.country:
+            raise VATException('Unable to get VAT charge because buyer country code is not specified.')
+
+        seller_country = settings.NODECONDUCTOR.get('SELLER_COUNTRY_CODE')
+        if not seller_country:
+            raise VATException('Unable to get VAT charge because seller country code is not specified.')
+
+        return pyvat.get_sale_vat_charge(
+            datetime.date.today(),
+            pyvat.ItemType.generic_electronic_service,
+            pyvat.Party(self.country, self.is_company and self.vat_code),
+            pyvat.Party(seller_country, True)
+        )
 
 
 @python_2_unicode_compatible
@@ -117,9 +187,6 @@ class Customer(core_models.UuidMixin,
 
     billing_backend_id = models.CharField(max_length=255, blank=True)
     balance = models.DecimalField(max_digits=9, decimal_places=3, null=True, blank=True)
-
-    is_company = models.BooleanField(default=False, help_text="Is company or private person")
-    country = core_fields.CountryField(blank=True)
 
     GLOBAL_COUNT_QUOTA_NAME = 'nc_global_customer_count'
 
@@ -328,7 +395,7 @@ class Project(core_models.DescribableMixin,
               core_models.NameMixin,
               core_models.DescendantMixin,
               quotas_models.QuotaModelMixin,
-              LoggableMixin,
+              StructureLoggableMixin,
               TimeStampedModel,
               StructureModel):
     class Permissions(object):
@@ -452,10 +519,6 @@ class Project(core_models.DescribableMixin,
     def get_log_fields(self):
         return ('uuid', 'customer', 'name', 'project_group')
 
-    @classmethod
-    def get_permitted_objects_uuids(cls, user):
-        return {'project_uuid': filter_queryset_for_user(cls.objects.all(), user).values_list('uuid', flat=True)}
-
     def get_parents(self):
         return [self.customer]
 
@@ -496,7 +559,7 @@ class ProjectGroup(core_models.UuidMixin,
                    core_models.NameMixin,
                    core_models.DescendantMixin,
                    quotas_models.QuotaModelMixin,
-                   LoggableMixin,
+                   StructureLoggableMixin,
                    TimeStampedModel):
     """
     Project groups are means to organize customer's projects into arbitrary sets.
@@ -575,16 +638,13 @@ class ProjectGroup(core_models.UuidMixin,
     def get_parents(self):
         return [self.customer]
 
-    @classmethod
-    def get_permitted_objects_uuids(cls, user):
-        return {'project_group_uuid': filter_queryset_for_user(cls.objects.all(), user).values_list('uuid', flat=True)}
-
 
 @python_2_unicode_compatible
 class ServiceSettings(quotas_models.ExtendableQuotaModelMixin,
                       core_models.UuidMixin,
                       core_models.NameMixin,
                       core_models.StateMixin,
+                      TagMixin,
                       LoggableMixin):
 
     class Meta:
@@ -605,7 +665,6 @@ class ServiceSettings(quotas_models.ExtendableQuotaModelMixin,
     options = JSONField(default={}, help_text='Extra options', blank=True)
     shared = models.BooleanField(default=False, help_text='Anybody can use it')
 
-    tags = TaggableManager(related_name='+', blank=True)
     tracker = FieldTracker()
 
     # service settings scope - VM that contains service
@@ -1068,6 +1127,7 @@ class ResourceMixin(MonitoringModelMixin,
                     core_models.SerializableAbstractMixin,
                     core_models.DescendantMixin,
                     LoggableMixin,
+                    TagMixin,
                     TimeStampedModel,
                     StructureModel):
 
@@ -1086,8 +1146,6 @@ class ResourceMixin(MonitoringModelMixin,
 
     service_project_link = NotImplemented
     backend_id = models.CharField(max_length=255, blank=True)
-    tags = TaggableManager(related_name='+', blank=True)
-
     start_time = models.DateTimeField(blank=True, null=True)
 
     def get_backend(self, **kwargs):